refactor: extract shared logic into AbstractConnectingBlockPlacementRule
- Created `AbstractConnectingBlockPlacementRule` to centralize shared connection logic.
- Updated `FencePlacementRule` and `VerticalSlimBlockPlacementRule` to extend the new abstract class.
